Productize IL assembly diff primitive
Move IL Diff's assembly/member pairing and summary producer into ILInspector.Instructions as IlAssemblyDiff, leaving Markout/card rendering in IlDiffHarness. Add a product-level self-diff unit test and document the assembly producer boundary.
