+++ docs/design/il-diff-canonicalization.md
@@ -5,6 +5,13 @@
 amount of canonicalization, then projects producer-owned rows through
 `IlDiffPrinter`.
 
+`IlAssemblyDiff` is the assembly/member producer above that substrate. It owns
+method identity for pairing bodies, runs self-diff and pair-diff checks, and
+emits summary counts, failure buckets, operation-family buckets, and typed
+example diffs. Harnesses own rendering and Markout/card projection; they should
+consume `IlAssemblyDiffResult` rather than reimplementing method matching or
+bucket wording.
+
 The boundary is intentionally narrow: the diff answers "which decoded IL
 operations changed?" It does not claim source equivalence, semantic equivalence,
 or durable identity for every operand category.
